T-50, T-53, T-54, T-55, T-56, T-57, T-58: Add PortfolioMetricsService by merging branch feature/T-50-T-53-portfolio-metrics-service-setup

+++ src/main/java/net/ictcampus/semodul/anlagendashboard/portfoliometrics/UserAssetJdbcDao.java
@@ -42,7 +42,7 @@ public class UserAssetJdbcDao implements UserAssetDao {
 							rs.getInt("User_ID"),
 							rs.getInt("Asset_ID"),
 							rs.getInt("Broker_ID"),
-							rs.getInt("Quantity"),
+							rs.getDouble("Quantity"),
 							rs.getTimestamp("PurchasedAt").toLocalDateTime(),
 							null // SoldAt ist bei offenen Positionen immer null
 					));
