+++ src/main/java/net/ictcampus/semodul/anlagendashboard/portfoliometrics/UserAssetModel.java
@@ -1,6 +1,4 @@
 /**
- * Represents a entry in the table user_asset with details such as ID, user ID, asset ID,
- * broker ID, quantity, purchase and sale timestamps.
  */
 package net.ictcampus.semodul.anlagendashboard.portfoliometrics;
 
@@ -11,11 +9,11 @@ public class UserAssetModel {
 	private int userId;
 	private int assetId;
 	private int brokerId;
-	private int quantity;
+	private double quantity;
 	private LocalDateTime purchasedAt;
 	private LocalDateTime soldAt;
 
-	public UserAssetModel(int idUserAsset, int userId, int assetId, int brokerId, int quantity,
+	public UserAssetModel(int idUserAsset, int userId, int assetId, int brokerId, double quantity,
 						  LocalDateTime purchasedAt, LocalDateTime soldAt) {
 		this.idUserAsset = idUserAsset;
 		this.userId = userId;
@@ -42,7 +40,7 @@ public class UserAssetModel {
 		return brokerId;
 	}
 
-	public int getQuantity() {
+	public double getQuantity() {
 		return quantity;
 	}
